requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
pyyaml==6.0.1
apscheduler==3.10.4
aiosqlite==0.20.0
jinja2==3.1.4
